refactor: hook 폴더 구조 재배치
